ci: Allow rerunning after trimming pipeline too
Reads the pipeline.template.yaml file again, so has to be applied too
