+++ misc/python/materialize/ci_util/trim_pipeline.py
@@ -13,10 +13,27 @@ import argparse
 import subprocess
 import sys
 from pathlib import Path
+from typing import Any
 
 import yaml
 
 from materialize import MZ_ROOT, spawn
+
+
+def permit_rerunning_successful_steps(pipeline: Any) -> None:
+    def visit(step: Any) -> None:
+        step.setdefault("retry", {}).setdefault("manual", {}).setdefault(
+            "permit_on_passed", True
+        )
+
+    for config in pipeline["steps"]:
+        if "trigger" in config or "wait" in config or "block" in config:
+            continue
+        if "group" in config:
+            for inner_config in config.get("steps", []):
+                visit(inner_config)
+            continue
+        visit(config)
 
 
 def main() -> int:
@@ -55,6 +72,9 @@ def main() -> int:
                 step["steps"] = new_inner_steps
                 del step["key"]
                 new_steps.append(step)
+
+    permit_rerunning_successful_steps(pipeline)
+
     spawn.runv(
         ["buildkite-agent", "pipeline", "upload", "--replace"],
         stdin=yaml.dump(new_steps).encode(),
